Move logic to abstract parent
